test(state): parameterize the state suite by dialect (Phase 3)
Every behavioural assertion about a store moves into a dialect-parameterized
factory, so Phase 4's PGlite leg runs the *identical* test bodies rather than a
hand-maintained subset that silently diverges.

`tests/state/store-suite.ts` exports `runStateDbSuite(makeDb, { dialect })` over
nine per-store modules in `tests/state/suites/`;
`tests/connectors/messaging/session-manager-suite.ts` is the second factory
(SessionManager is not reachable through StateDb, so it takes its own ctx).
`tests/state/db.test.ts` and `session-manager.test.ts` become thin sqlite
runners.

Scope is wider than the phase doc's inventory table, which was written against a
3-store codebase: seven further pure-state suites moved in (cron-run-store,
repo-ref's qualifiedRepoSql, execution-outcome, execution-store-reads,
feedback-store, team-store, approval-store). They are the only guards for the
ports 00-architecture flags as the highest PG risk — `rowid`, `strposExpr`,
`sumTrue`/`sumFalse`, the ON CONFLICT upserts, `INSERT OR IGNORE`. Left out, the
PG leg could go green proving none of them.

Not parameterizable, so placed rather than moved: `isTriggerActorType` (pure
predicate) to its own file, repo-ref's pure half left behind, the boot compat
step's owner backfill to repo-normalization.test.ts, and the legacy messaging
rebuild to session-manager.legacy.test.ts.

Assertion changes are only the four sanctioned ports: the rotted 2026-04-*
dailyStats literals (which had degraded to asserting zero-filled rows are
sorted) are relative dates again and the case is strengthened; `daysAgo` slices
its bucket key off the string it inserts; the partial-unique-index probe matches
`isUniqueViolation` instead of /UNIQUE/; the anchor row-count peek narrows to a
text column.

Found while moving, and recorded as a Phase 4 blocker: `asStateClient()` cannot
carry the PG leg on its own. The query-builder surface is structurally identical
across drivers, but per-column value mapping is not and does not go through
dialect.ts — a sqliteTable used on a PG client sends 1 into a boolean and
JSON.parses an already-parsed jsonb. Booleans break on write, JSON on read, and
every store imports schema/sqlite.js directly. Verified against drizzle-orm
0.45.2 source; 04 now opens with a warning block.

206 files / 3,127 tests -> 199 files / 3,128 tests (the specified +1, nothing
else). Zero src/ diff, zero dialect skips.
